client search query to shared concern
