+++ app/models/grda_warehouse/client_search_query.rb
@@ -9,13 +9,11 @@ module GrdaWarehouse
   # - Fingerprinting based on query content to allow finding existing searches
   # - Uses UUID primary key for secure URL sharing
   class ClientSearchQuery < GrdaWarehouseBase
+    include ClientSearchQueryShared
     belongs_to :created_by, class_name: 'User'
 
-    MAX_STRING_LENGTH = 100
     ALLOWED_PARAMS = ['q', 'client'].freeze
     ALLOWED_CLIENT_PARAMS = ['first_name', 'last_name', 'dob', 'ssn'].freeze
-
-    validate :validate_params
 
     # @param params [ActionController::Parameters] request params
     # @return [ActionController::Parameters, nil] Permitted parameters or nil if no valid params present
@@ -25,24 +23,7 @@ module GrdaWarehouse
 
     def self.find_or_create_by_params(params, user:)
       norm = normalize_params(params.to_h)
-
-      # Validate params first
-      instance = new(params: norm)
-      instance.validate_params
-      return instance if instance.errors.any?
-
-      fingerprint = generate_fingerprint(norm)
-      upsert(
-        { fingerprint: fingerprint, params: norm, created_by_id: user.id },
-        unique_by: :fingerprint,
-        on_duplicate: Arel.sql('params = EXCLUDED.params, created_by_id = EXCLUDED.created_by_id'),
-      )
-
-      find_by!(fingerprint: fingerprint)
-    end
-
-    def self.generate_fingerprint(params)
-      Digest::SHA256.hexdigest(params.to_json)
+      find_or_create_by_normalized_params(norm, user: user)
     end
 
     def self.normalize_params(params)
@@ -58,35 +39,6 @@ module GrdaWarehouse
           v
         end
       end.reject { |_, v| v.blank? }.sort.to_h
-    end
-
-    def validate_params
-      return if params.blank?
-
-      # Validate top-level parameters
-      invalid_params = params.keys - ALLOWED_PARAMS
-      errors.add(:params, "contains invalid parameters: #{invalid_params.join(', ')}") if invalid_params.any?
-
-      # Validate client parameters if present
-      if params['client'].present?
-        invalid_client_params = params['client'].keys - ALLOWED_CLIENT_PARAMS
-        errors.add(:params, "contains invalid client parameters: #{invalid_client_params.join(', ')}") if invalid_client_params.any?
-      end
-
-      # Validate string lengths
-      validate_string_lengths(params)
-    end
-
-    def validate_string_lengths(hash, prefix = nil)
-      hash.each do |key, value|
-        case value
-        when String
-          field = prefix ? "#{prefix}.#{key}" : key
-          errors.add(:params, "#{field} is too long (max #{MAX_STRING_LENGTH} characters)") if value.length > MAX_STRING_LENGTH
-        when Hash
-          validate_string_lengths(value, prefix ? "#{prefix}.#{key}" : key)
-        end
-      end
     end
 
     def query_params
